orchestrator/requirements.txt
flask==2.3.3
flask-restx==1.2.0
requests==2.31.0
pyyaml==6.0.1
asyncio